* gnuradio/next: (806 commits)
  gruel: added missing ignores
  gruel: fixed swig interface file to dereference pmt_t.
  qtgui: fix distcheck error
  gruel: fixing structure. Passes make check.
  gruel: SWIGing Gruel into Python to access PMTs.
  gnuradio-examples: add C++ audio examples using new gr-audio
  created gruel/attributes.h to house compiler specific attribute macros
  audio: remove obsoleted individual top-level components
  gr-audio: added README and default config fix
  volk: simplify the get new method for the aligned pool
  grc: moved all usrp1 and usrp2 stuff out of grc and into gr-usrp*/grc
  grc: swap store the subprocess object rather than the pid when executing
  qtgui: removed python directory that was added, never used
  uhd: use %ignore to hide warnings and fix errors
  Added/updated ignore files.
  Fixing gr_filter_design program to import from gnuradio Python package.
  audio: high prio for platform specific audio osx
  audio: added windows and osx audio source files
  audio: added config checks for other audios, added jack and port
  audio: make prefs look like old audio, removed old audio.py
  ...

+++ gnuradio-core/src/lib/filter/gri_fft_filter_fff_generic.cc
@@ -86,7 +86,7 @@ gri_fft_filter_fff_generic::compute_sizes(int ntaps)
 {
   int old_fftsize = d_fftsize;
   d_ntaps = ntaps;
-  d_fftsize = (int) (2 * pow(2.0, ceil(log(ntaps) / log(2))));
+  d_fftsize = (int) (2 * pow(2.0, ceil(log(double(ntaps)) / log(2.0))));
   d_nsamples = d_fftsize - d_ntaps + 1;
 
   if (0)
